Rackspace Cloud Block Storage models have the potential to create unintended resources

+++ tests/rackspace/models/block_storage/volume_tests.rb
@@ -6,7 +6,11 @@ Shindo.tests('Fog::Rackspace::BlockStorage | volume', ['rackspace']) do
   options = { :display_name => "fog_#{Time.now.to_i.to_s}", :size => 100 }
 
   model_tests(service.volumes, options, false) do
-    @instance.wait_for { ready? }
+    @instance.wait_for(timeout=1200) { ready? }
+
+    tests('double save').raises(Fog::Rackspace::BlockStorage::IdentifierTaken) do
+      @instance.save
+    end
 
     tests('#attached?').succeeds do
       @instance.state = 'in-use'
@@ -20,7 +24,7 @@ Shindo.tests('Fog::Rackspace::BlockStorage | volume', ['rackspace']) do
 
         returns(true) { @instance.snapshots.first.id == snapshot.id }
       ensure
-        snapshot.destroy
+        snapshot.destroy if snapshot
       end
     end
